Frontend fixes
@Rafipilot please try this branch locally and let me know what you think. It's easier for us to test this way, with limited genres/search terms, but still not getting the results we'd expect regularly.

I suspect it's a result of how we're pre-training the agent. It may be better to, instead of pretraining, withhold agent recommendations for the first 5 videos, to get 5 button clicks from the user first before we start using the agent's recommendations (this is the "cold-start" problem as its called in recsys). I'll try doing that tomorrow.

I'll delay pushing these changes to the hosted app (let's call it production) until you review.

Co-Authored-By: Rafayel Latif <74052861+Rafipilot@users.noreply.github.com>